Added AcctThread.py utility:
AcctThread is a command line utility that takes a ripple account
ID as an argument then walks the PreviousTxnID thread left
behind in the account root.  This walk will, if enough history
is available, pass through all transactions affecting the
account root in question until we arrive at the creation of the
account root.

Output is JSON (with text separators and text comments) to
stdout.  To capture the output to a file redirect the output.